Better project directory layout
- Follow recommendations for organizing medium/large projects
