+++ src/kafkaaggregator/testtopic/models.py
